+++ packages/eslint-plugin/src/rules/no-unsafe-assignment.ts
@@ -2,8 +2,10 @@ import {
   TSESTree,
   AST_NODE_TYPES,
 } from '@typescript-eslint/experimental-utils';
+import * as tsutils from 'tsutils';
 import * as ts from 'typescript';
 import * as util from '../util';
+import { getThisExpression } from '../util';
 
 const enum ComparisonType {
   /** Do no assignment comparison */
@@ -25,13 +27,17 @@ export default util.createRule({
       requiresTypeChecking: true,
     },
     messages: {
-      anyAssignment: 'Unsafe assignment of an any value.',
-      unsafeArrayPattern: 'Unsafe array destructuring of an any array value.',
+      anyAssignment: 'Unsafe assignment of an `any` value.',
+      anyAssignmentThis: [
+        'Unsafe assignment of an `any` value. `this` is typed as `any`.',
+        'You can try to fix this by turning on the `noImplicitThis` compiler option, or adding a `this` parameter to the function.',
+      ].join('\n'),
+      unsafeArrayPattern: 'Unsafe array destructuring of an `any` array value.',
       unsafeArrayPatternFromTuple:
-        'Unsafe array destructuring of a tuple element with an any value.',
+        'Unsafe array destructuring of a tuple element with an `any` value.',
       unsafeAssignment:
         'Unsafe assignment of type {{sender}} to a variable of type {{receiver}}.',
-      unsafeArraySpread: 'Unsafe spread of an any value in an array.',
+      unsafeArraySpread: 'Unsafe spread of an `any` value in an array.',
     },
     schema: [],
   },
@@ -39,6 +45,11 @@ export default util.createRule({
   create(context) {
     const { program, esTreeNodeToTSNodeMap } = util.getParserServices(context);
     const checker = program.getTypeChecker();
+    const compilerOptions = program.getCompilerOptions();
+    const isNoImplicitThis = tsutils.isStrictCompilerOptionEnabled(
+      compilerOptions,
+      'noImplicitThis',
+    );
 
     // returns true if the assignment reported
     function checkArrayDestructureHelper(
@@ -243,9 +254,27 @@ export default util.createRule({
           return false;
         }
 
+        let messageId: 'anyAssignment' | 'anyAssignmentThis' = 'anyAssignment';
+
+        if (!isNoImplicitThis) {
+          // `var foo = this`
+          const thisExpression = getThisExpression(senderNode);
+          if (
+            thisExpression &&
+            util.isTypeAnyType(
+              util.getConstrainedTypeAtLocation(
+                checker,
+                esTreeNodeToTSNodeMap.get(thisExpression),
+              ),
+            )
+          ) {
+            messageId = 'anyAssignmentThis';
+          }
+        }
+
         context.report({
           node: reportingNode,
-          messageId: 'anyAssignment',
+          messageId,
         });
         return true;
       }