+++ packages/eslint-plugin/src/util/index.ts
@@ -4,6 +4,7 @@ export * from './astUtils';
 export * from './collectUnusedVariables';
 export * from './createRule';
 export * from './getFunctionHeadLoc';
+export * from './getThisExpression';
 export * from './getWrappingFixer';
 export * from './isTypeReadonly';
 export * from './misc';